Adapt packages to use RunProtocol type class
The way ouroboros-network exposes Protocol information to clients such
as cardano-api and cardano-node was changed from a Protocol GADT to a
type class called RunProtocol; similarly for ProtocolClient and
RunProtocolClient. The motivation behind this change is to make it
easier to create new consensus modes for research and prototype purposes
and integrate them into the cardano infrastructure.

The protocol GADT was a closed world type, requiring
ouroboros-network to list somewhere all of the consensus modes
implemented there for client consumption even if they are logically
independent. The type class approach allows consensus modes to expose
information to client code without knowing about each other.

cardano-api, cardano-cli, cardano-node, and cardano-node-chairman are
here modified to use this type class as much as possible. At some point
they do have to close their assumptions on which modes are supported but
this is pushed up to the surface as much as possible.