Fix PrototypeMode hardfork versions
The protocol versions at which the network forks to Prototype One and Two are specified in Cardano.Node.Protocol.Voltaire. Previously, this reused the logic from the Cardano mode, which specified these versions in a configuration file, and otherwise defaults to v3.0 and v4.0, respectively. Since our Shelley era is v0.0, and it's only allowed to increase one major version at a time, it wasn't possible to upgrade the network to PrototypeOne. Fix this by specifying v1.0 and v2.0 as the versions at which the network upgrades to PrototypeOne and PrototypeTwo, respectively. Also no need to take the CardanoMode-related options as argument, as we don't use them in the PrototypeMode.