add CLI commands: create-mir-proposal, mir-proposal-file
The "create-mir-proposal" subcommand lives under the "governance" command. It writes a MIR proposal to a file (similar to "create-update-proposal").
The "mir-proposal-file" subcommand lives under the "transaction build-raw" command. It reads a MIR proposal from a file and includes it in a transaction (similar to "update-proposal-file").

Limitation: it's not possible to create a transaction that contains _both_ a protocol parameter update proposal and a MIR proposal.